# Simplify computation of return type in broadcast