Since we rely on inference, we can use `_return_type` directly instead of via a complex machinery.

As suggested by @mbauman at https://github.com/JuliaLang/julia/pull/39185#issuecomment-758142273.