@timholy 2021-01-17
Does this have an impact on the inference & codegen time? The broadcast infrastructure is already a big piece of the latency for many packages, just curious whether this makes it better or worse.