@nalimilan 2021-01-17
Here's a small benchmark with `x = rand(10)`, each time in a fresh Julia session.

Master:
```julia
julia> @time exp.(x);
  0.071312 seconds (207.54 k allocations: 12.900 MiB, 99.55% compilation time)

julia> @time exp.(x);
  0.073889 seconds (207.54 k allocations: 12.900 MiB, 99.48% compilation time)

julia> @time exp.(x);
  0.072427 seconds (207.54 k allocations: 12.900 MiB, 99.54% compilation time)
```

PR:
```julia
julia> @time exp.(x);
  0.075400 seconds (223.03 k allocations: 13.804 MiB, 99.46% compilation time)

julia> @time exp.(x);
  0.071174 seconds (223.03 k allocations: 13.804 MiB, 99.56% compilation time)

julia> @time exp.(x);
  0.077204 seconds (223.03 k allocations: 13.804 MiB, 99.58% compilation time)
```

So there are a few more allocations, and it might be a bit slower, but it's not super clear. Do you have ideas about other possible benchmarks?