@timholy 2021-01-17
Maybe one where `f` has multiple arguments? As long as that looks good too, then I'm fine with this idea.