@nalimilan 2021-01-17
Here's what I get for slightly more complex cases (still with a fresh session for each pair of commands):
master:
```julia
julia> @time x .+ 1;
  0.062047 seconds (165.10 k allocations: 10.133 MiB, 99.52% compilation time)

julia> @time Float32.(x) .+ x .+ 1;
  0.132495 seconds (281.68 k allocations: 16.544 MiB, 99.32% compilation time)

julia> @time x .+ 1;
  0.063509 seconds (165.10 k allocations: 10.133 MiB, 99.48% compilation time)

julia> @time Float32.(x) .+ x .+ 1;
  0.137296 seconds (281.68 k allocations: 16.544 MiB, 99.36% compilation time)

julia> @time x .+ 1;
  0.061300 seconds (165.10 k allocations: 10.133 MiB, 99.34% compilation time)

julia> @time Float32.(x) .+ x .+ 1;
  0.136219 seconds (281.68 k allocations: 16.544 MiB, 99.40% compilation time)
```


PR:
```julia
julia> @time x .+ 1;
  0.065680 seconds (180.26 k allocations: 11.009 MiB, 99.58% compilation time)

julia> @time Float32.(x) .+ x .+ 1;
  0.134870 seconds (296.66 k allocations: 17.251 MiB, 99.40% compilation time)

julia> @time x .+ 1;
  0.065040 seconds (180.26 k allocations: 11.009 MiB, 99.35% compilation time)

julia> @time Float32.(x) .+ x .+ 1;
  0.141318 seconds (296.66 k allocations: 17.251 MiB, 99.39% compilation time)

julia> @time x .+ 1;
  0.068847 seconds (180.26 k allocations: 11.009 MiB, 99.54% compilation time)

julia> @time Float32.(x) .+ x .+ 1;
  0.135004 seconds (296.66 k allocations: 17.251 MiB, 99.32% compilation time)
```

So still a slight increase in allocations.

But I've found a more serious problem: the CI failure is due to `combine_eltypes` being used at
https://github.com/JuliaLang/julia/blob/26a721b28ad56c006957ee1f2de083befcfe7904/stdlib/SparseArrays/src/higherorderfns.jl#L163
and
https://github.com/JuliaLang/julia/blob/26a721b28ad56c006957ee1f2de083befcfe7904/stdlib/SparseArrays/src/higherorderfns.jl#L191
I'm not sure we can actually get rid of these without reinventing most of `combine_eltypes`. What do you think? BTW, I'm surprised that `combine_type` is used to determine the type of the result (even when not empty) as it relies on inference.