@vtjnash 2021-01-18
SparseArrays may have some legacy issues with the way it forms the eltype. I think this PR seems reasonable. Should be similar time, since we're about to infer in to the methods (for the runtime code path) anyways.