@mbauman 2021-01-20
> Do you have ideas about other possible benchmarks?

The tests themselves tend to lend themselves fairly nicely to compile-time benchmarking.  E.g., `time julia test/broadcast.jl` or some such.