@vtjnash 2021-08-18
Note the currently open question of whether this is actually better or worse (https://github.com/JuliaLang/julia/pull/39295#discussion_r615853892)