@nalimilan 2021-12-29
> Can we wake up this? My local bench shows that this PR is able to reduce about 10% ~ 13% of the time cost of `Base.runtests("broadcast")`

The time it takes to run tests isn't usually a very interesting benchmark as tests are a very atypical coding pattern. Do you have evidence that this PR improves performance (or compile times) on real use cases? This isn't to say that I'm opposed to merging it.