@N5N3 2021-12-29
Well l have no further envidence, I just follows above advice from @mbauman to bench.
And found that the time cost and mem usage reduced after similar commit.
IIRC, we also use the test itself to bench the codegen improvement from avoiding always inline.
Maybe we need a package whose TTFP is dominated by broadcast?