@vtjnash 2022-01-07
The PR is currently wrong, though I have a suggestion above to fix it. The remaining question however, as before, is whether we want this design change (https://github.com/JuliaLang/julia/pull/39295#discussion_r615853892)